requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
appnope==0.1.4
asgiref==3.11.0
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
bitarray==3.4.2
certifi==2025.4.26
charset-normalizer==3.4.2
ckzg==2.1.1
comm==0.2.2
cytoolz==1.0.1
debugpy==1.8.14
decorator==5.2.1
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
executing==2.2.0
frozenlist==1.6.0
hexbytes==1.3.1
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
multidict==6.4.4
mysql-connector-python==9.4.0
nest-asyncio==1.6.0
packaging==23.2
parsimonious==0.10.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-solc-x==2.0.3
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pyunormalize==16.0.0
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tkcalendar==1.6.1
toolz==1.0.0
tornado==6.4.2
traitlets==5.14.3
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
web3==7.12.0
websockets==15.0.1
yarl==1.20.0
